html/semantics/forms/textfieldselection/textfieldselection-setRangeText.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>text field selection: setRangeText</title>
<link rel="author" title="Denis Ah-Kang" href="mailto:denis@w3.org">
<link rel=help href="http://www.w3.org/html/wg/drafts/html/master/#textFieldSelection">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<input value="foobar">
<textarea>foobar</textarea>
<script>
  var input = document.querySelector("input"),
      textarea = document.querySelector("textarea");

  ["input", "textarea"].forEach(function(el) {
    var element = document.querySelector(el),
        t = async_test(el + ".setRangeText fires a select event");

    element.onselect = t.step_func_done(function(e) {
        assert_true(e.isTrusted, "event is trusted");
        assert_false(e.bubbles, "event bubbles");
        assert_false(e.cancelable, "event is not cancelable");
    });

    test(function() {
      assert_equals(element.value, "foobar");
      element.selectionStart = 0;
      element.selectionEnd = 3;
      assert_equals(element.selectionStart, 0);
      assert_equals(element.selectionEnd, 3);
      element.setRangeText("foobar2");
      assert_equals(element.value, "foobar2bar");
      assert_equals(element.selectionStart, 0);
      assert_equals(element.selectionEnd, 7);
      element.setRangeText("foobar3", 7, 10);
      assert_equals(element.value, "foobar2foobar3");
    }, el + ".setRangeText with only one argument replaces the value between selectionStart and selectionEnd, otherwise replaces the value between 2nd and 3rd arguments");

    test(function(){
      element.value = "foobar";
      element.selectionStart = 0;
      element.selectionEnd = 0;

      element.setRangeText("foobar2", 0, 3); // no 4th arg, default "preserve"
      assert_equals(element.value, "foobar2bar");
      assert_equals(element.selectionStart, 0);
      assert_equals(element.selectionEnd, 0);
    }, el + " selectionMode missing");

    test(function(){
      element.setRangeText("foo", 1, 7, "select");
      assert_equals(element.value, "ffoobar");
      assert_equals(element.selectionStart, 1);
      assert_equals(element.selectionEnd, 4);
    }, el + " selectionMode 'select'");

    test(function(){
      element.setRangeText("oo", 1, 4, "start");
      assert_equals(element.value, "foobar");
      assert_equals(element.selectionStart, 1);
      assert_equals(element.selectionEnd, 1);
    }, el + " selectionMode 'start'");

    test(function(){
      element.setRangeText("foobar", 3, 6, "end");
      assert_equals(element.value, "foofoobar");
      assert_equals(element.selectionStart, 9);
      assert_equals(element.selectionEnd, 9);
    }, el + " selectionMode 'end'");

    test(function(){
      element.setRangeText("", 3, 6, "preserve");
      assert_equals(element.value, "foobar");
      assert_equals(element.selectionStart, 6);
      assert_equals(element.selectionEnd, 6);
    }, el + " selectionMode 'preserve'");

    test(function(){
      assert_throws("INDEX_SIZE_ERR", function() {
        element.setRangeText("barfoo", 2, 1);
      });
    }, el + ".setRangeText with 3rd argument greater than 2nd argument throws an IndexSizeError exception");
  })
</script>
